+++ examples/balls_vee3d.rs
@@ -31,24 +31,24 @@ fn main() {
     GraphicsManager::simulate(balls_vee_3d)
 }
 
-pub fn balls_vee_3d(window: &mut Window, graphics: &mut GraphicsManager) -> dim3::BodyWorld3d<f64> {
+pub fn balls_vee_3d(window: &mut Window, graphics: &mut GraphicsManager) -> dim3::BodyWorld3d<f32> {
     /*
      * World
      */
     let mut world = BodyWorld::new();
-    world.set_gravity(Vec3::new(0.0f64, -9.81, 0.0));
+    world.set_gravity(Vec3::new(0.0f32, -9.81, 0.0));
 
     /*
      * Planes
      */
     let normals = [
-        Vec3::new(-1.0f64, 1.0, -1.0 ),
-        Vec3::new(1.0f64, 1.0, -1.0 ),
-        Vec3::new(-1.0f64, 1.0, 1.0 ),
-        Vec3::new(1.0f64, 1.0, 1.0 )
+        Vec3::new(-1.0f32, 1.0, -1.0 ),
+        Vec3::new(1.0f32, 1.0, -1.0 ),
+        Vec3::new(-1.0f32, 1.0, 1.0 ),
+        Vec3::new(1.0f32, 1.0, 1.0 )
     ];
     for n in normals.iter() {
-        let rb   = RigidBody::new(Geom::new_plane(Plane::new(*n)), 0.0f64, Static, 0.3, 0.6);
+        let rb   = RigidBody::new(Geom::new_plane(Plane::new(*n)), 0.0f32, Static, 0.3, 0.6);
         let body = @mut RB(rb);
 
         world.add_body(body);
@@ -58,20 +58,20 @@ pub fn balls_vee_3d(window: &mut Window, graphics: &mut GraphicsManager) -> dim3
     /*
      * Create the balls
      */
-    let num     = (1500.0f64.pow(&(1.0f64 / 3.0))) as uint;
+    let num     = (1500.0f32.pow(&(1.0f32 / 3.0))) as uint;
     let rad     = 0.5;
     let shift   = 2.5 * rad;
-    let centerx = shift * (num as f64) / 2.0;
-    let centery = shift * (num as f64) / 2.0;
+    let centerx = shift * (num as f32) / 2.0;
+    let centery = shift * (num as f32) / 2.0;
 
     for i in range(0u, num) {
         for j in range(0u, num) {
             for k in range(0u, num) {
-                let x = i as f64 * 2.5 * rad - centerx;
-                let y = 10.0 + j as f64 * 2.5 * rad + centery * 2.0;
-                let z = k as f64 * 2.5 * rad - centerx;
+                let x = i as f32 * 2.5 * rad - centerx;
+                let y = 10.0 + j as f32 * 2.5 * rad + centery * 2.0;
+                let z = k as f32 * 2.5 * rad - centerx;
 
-                let mut rb = RigidBody::new(Geom::new_ball(Ball::new(rad)), 1.0f64, Dynamic, 0.3, 0.6);
+                let mut rb = RigidBody::new(Geom::new_ball(Ball::new(rad)), 1.0f32, Dynamic, 0.3, 0.6);
 
                 rb.append_translation(&Vec3::new(x, y, z));
