+++ examples/boxes_vee2d.rs
@@ -29,18 +29,18 @@ fn main() {
     GraphicsManager::simulate(boxes_vee_2d)
 }
 
-pub fn boxes_vee_2d(graphics: &mut GraphicsManager) -> dim2::BodyWorld2d<f64> {
+pub fn boxes_vee_2d(graphics: &mut GraphicsManager) -> dim2::BodyWorld2d<f32> {
     /*
      * World
      */
     let mut world = BodyWorld::new();
-    world.set_gravity(Vec2::new(0.0f64, 9.81));
+    world.set_gravity(Vec2::new(0.0f32, 9.81));
 
     /*
      * First plane
      */
-    let geom   = Plane::new(Vec2::new(-1.0f64, -1.0));
-    let mut rb = RigidBody::new(Geom::new_plane(geom), 0.0f64, Static, 0.3, 0.6);
+    let geom   = Plane::new(Vec2::new(-1.0f32, -1.0));
+    let mut rb = RigidBody::new(Geom::new_plane(geom), 0.0f32, Static, 0.3, 0.6);
 
     rb.append_translation(&Vec2::new(0.0, 10.0));
 
@@ -52,8 +52,8 @@ pub fn boxes_vee_2d(graphics: &mut GraphicsManager) -> dim2::BodyWorld2d<f64> {
     /*
      * Second plane
      */
-    let geom   = Plane::new(Vec2::new(1.0f64, -1.0));
-    let mut rb = RigidBody::new(Geom::new_plane(geom), 0.0f64, Static, 0.3, 0.6);
+    let geom   = Plane::new(Vec2::new(1.0f32, -1.0));
+    let mut rb = RigidBody::new(Geom::new_plane(geom), 0.0f32, Static, 0.3, 0.6);
 
     rb.append_translation(&Vec2::new(0.0, 10.0));
 
@@ -65,20 +65,20 @@ pub fn boxes_vee_2d(graphics: &mut GraphicsManager) -> dim2::BodyWorld2d<f64> {
     /*
      * Create the boxes
      */
-    let num = (1000.0f64.sqrt()) as uint;
+    let num = (1000.0f32.sqrt()) as uint;
     let rad = 0.5;
     let shift   = 2.5 * rad;
-    let centerx = shift * (num as f64) / 2.0;
-    let centery = shift * (num as f64) / 2.0;
+    let centerx = shift * (num as f32) / 2.0;
+    let centery = shift * (num as f32) / 2.0;
 
     for i in range(0u, num) {
         for j in range(0u, num) {
-            let x = i as f64 * 2.5 * rad - centerx;
-            let y = j as f64 * 2.5 * rad - centery * 2.0 - 10.0;
+            let x = i as f32 * 2.5 * rad - centerx;
+            let y = j as f32 * 2.5 * rad - centery * 2.0 - 10.0;
 
             let box    = Box::new(Vec2::new(rad, rad));
             let geom   = Geom::new_box(box);
-            let mut rb = RigidBody::new(geom, 1.0f64, Dynamic, 0.3, 0.6);
+            let mut rb = RigidBody::new(geom, 1.0f32, Dynamic, 0.3, 0.6);
 
             rb.append_translation(&Vec2::new(x, y));
